Add files via upload
So, in this paper, we will propose a data summary tables, visualizations which make some comparisons by using different properties of participants, and a deep learning model which first train the model by using the experiment outcomes analysis dataset and then test the accuracy, so that we can understand how much use the collected data is. Visualizations will be provided by using both R and Python and their rich visualization tools libraries. These visualizations consist of comparisons between the participants who are working in a silent environment and noise environment.